fix(release): pin consolidated artifact identity
Bind the gate to the exact source tree, tarball, package file hashes, bundle/map debug ID and mapped checkout sources; move the checker into the typed source tree and add adversarial coverage.
